requirements.txt
absl-py==0.12.0
async-generator==1.10
attrs==21.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
cycler==0.10.0
defusedxml==0.7.1
entrypoints==0.3
flake8==3.7.9
flake8-bugbear==20.1.4
flake8-comprehensions==3.2.2
flake8-mypy==17.8.0
flake8-pyi==19.3.0
Flask==1.1.2
future==0.18.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
-e git+https://github.com/NVIDIAGameWorks/kaolin@5554bce48187d2acfcbce62cc5f8c511a082de92#egg=kaolin
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
more-itertools==8.8.0
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.20.2
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.54
packaging==20.9
pandocfilters==1.4.3
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.17.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.2
pytest-cov==2.8.1
python-dateutil==2.8.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
testpath==0.5.0
toml==0.10.2
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.61.0
traitlets==5.0.5
trimesh==3.9.20
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
usd-core==20.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
zipp==3.4.1
